Add files via upload
Description:
This project is a Python-based application that performs sentiment analysis on movie reviews. It scrapes user reviews from the web, cleans the data, and analyzes the sentiment using the Afinn library. The project includes a Tkinter-based GUI that allows users to input a movie name and view the results of the sentiment analysis, categorized into positive, negative, and neutral reviews.